requirements.txt
discord.py
py-term
cryptography
Pillow
Flask
Flask-SocketIO
aiohttp
tmdbv3api